Fixed a bug in copy in which the type of a parameter symbol was copied when a parameter symbol was copied and the return type of a function was copied when the function definition statement was copied.  These types are now shared, copying them led to an infinite recursion.
Quick-fixed with-statement application so that it can be applied
before resolving symbols by having it resolve the class that is
applied via the with-expression.

Moved the with-statement application traversal first in the suite of
Cleanup traversals.

Moved test_class_with2.chpl out of the future directory.  This test
tests with on classes when the classes have methods.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@1107 3a8e244f-b0f2-452b-bcba-4c88e055c3ca