Moved resolution of ParenOpExpr function symbols after analysis. Previously only resolution of MemberAccess symbols was done after analysis.  In this post-analysis resolution, analysis's call_info function is called to determine which function to resolve to.  This supports overloading of functions.  For example, the test test_overload_function1.chpl now works.
Moved around a few test cases motivating type variables,
instantiation, and cloning.  These are now in a future directory,
though the overloading test is set to run.

Made a slight change to call_info, directed by John, so that the
interface now only requires a ParenOpExpr, not a ParenOpExpr and the
containing function.  This is because the containing function can be
determined from the ParenOpExpr (it's statement's parentSymbol (which
is hopefullly a function and not a class--no function calls in class
definitions)).

This checkin breaks one test:
[Error matching compiler output for trivial/deitz/test_write_return]
Here the problem is that we're not resolving this function until after
analysis.  Still, analysis should be able to give us the type?  I'm
going to look for this bug, but I'm checking in for help.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@1119 3a8e244f-b0f2-452b-bcba-4c88e055c3ca