Added rudimentary support for unions (without runtime typechecking) from parsing to code generation.  The single test case currently fails in analysis, but I'm not sure why.  Expect a regression: classes/deitz/test_union1.chpl.
Unions are now generated as structs with an id tag (for runtime
typechecking) and a union for efficient storage of the union fields.
In the AST, the union is a class just like a record.  In the case of
the record, a bool flag value is set to true.  Similarly, for unions,
a bool flag union_value is set set to true.

Fixed a bug in the setup of nilClassType in which uninitialized
nilClassType was passed to its own constructor.  This was coerced to
false, but now that there are two parameters required by this
constructor, it no longer worked.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@1127 3a8e244f-b0f2-452b-bcba-4c88e055c3ca