Factored copy in Stmt, Expr, Symbol, Type, and their subclasses into two pieces like traverse.  There is now an external copy that should be used, along with copyList, and internal copyStmt, etc., that does the actual copying.
Added a callback to analysis for cloning functions so that the ainfo
part of the BaseAST can be maintained in clone functions.  This uses a
new class called CloneCallback.  It is passed into copy with a default
value of NULL.  In the NULL case, nothing happens.  If it's not NULL,
then the external factor of copy calls back into analysis to update
the ainfo.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@1172 3a8e244f-b0f2-452b-bcba-4c88e055c3ca