Removed Symboltable::copyParams and handle the copying of parameters in the copying of a FnDefStmt.  The parameters are now copied with the CloneCallback for the case of a function being called.  This uses copyList rather than copyParams which basically did the same thing.
Added a guard against copying the NilSymbol so that it is just
returned.

Changed the error message in codegen having to do with unresolved
types.  If the type of an expression is unknown at codegen time, we
now say "unknown type encountered in codegen" rather than " shouldn't
be calling into analysis now".


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@1204 3a8e244f-b0f2-452b-bcba-4c88e055c3ca