Added a few tests, three new futures.
The test test_method7 calls an overloaded function from within a
method.  It currently works.

The overload_integerfloat tests test overloading foo(i, i), foo(i, f),
and foo(f, f) where f is a float and i is an integer.  One problem
with these is that when a float is initialized with an integer value,
it is taken to be an integer by analysis even though it should be
taken to be a float.  The declared type should override the inferred
type from initialization.  Right?

The second problem is that when foo is called as in foo(f, i), it
fails to resolve even though it should resolve to foo(f, f).  Right?


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@1244 3a8e244f-b0f2-452b-bcba-4c88e055c3ca