Expected regression! This checkin breaks the following test(s):   trivial/bradc/protoast/mg-seq-concrete See note(s) below.
Changed the mechanism through which symbols are defined in the
symboltable.  Previously, the creation of symbols was decoupled from
defining them in the symboltable.  They are now defined in the
symboltable when they are constructed.  This greatly simplifies many
aspects of the compiler.

Made a new scope called nilsScope for declaring nilSymbol,
nilFnSymbol, and nilVarSymbol.

Implemented copy for VarSymbol and TypeSymbol.  The others are already
implemented.

Added a new procedure "TypeDefStmt* TypeDefStmt::clone(CloneCallback*
clone_callback)" for cloning types such as classes.  This is ready
(pending bugs) to be called by analysis in order to clone classes with
underspecified types, e.g., a class with a variable field with no
type.  Handed the futures of
test/types/type_variables/deitz/test_clone_class[1|2] to John.

Fixed a small bug in which the back pointer was not set for
ForallExprs that were modified by setForallExpr.

Partially cleaned up how unions are handled.  The enums for unions
should eventually be made statements of their own rather than be
handled in code generation.  Now at least the symbols are added to the
symboltables.  The InsertUnionChecks looks up these symbols rather
than creating but not defining its own.

Changed how parameters to procedures are parsed.  Commas, rather than
semicolons, are now used between intent-variable-type formals.
Multiple variables will eventually be allowed with blank space.  This
change breaks mg-seq-concrete.

Removed long-dead traversal DefineSymbols.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@1269 3a8e244f-b0f2-452b-bcba-4c88e055c3ca