This is an up-for-air checkin.  It greatly increases the stability of copying things in the AST though more work is still needed.
Added a new fixup verification called using fixup "hyper verify".
This does the regular fixup "verify" but also copies and replaces
every statement and expression in the AST.  The point of "hyper
verify" is to catch bugs in copy, replace, and elsewhere.  I am
finding it hugely useful.  Though I have a number of things left to
fix (that are caught when "hyper verify" is used), this checkin fixes
many things and is stable without the "hyper verify" pass.

The copy interface is undergoing major changes.  These will be
documented when they are finished.  The basic idea is to get copying
working well either before or after analysis and when the symbols and
types should or should not be copied.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@1292 3a8e244f-b0f2-452b-bcba-4c88e055c3ca