Got parameter passing intents for integers working (finally!). Modified Steve's new copy() protocol so that when his clone flag is false, Variable Expr's Symbols are not changed into UnresolvedSymbols, but merely re-used as-is.  In my case, I just want to make a trivial copy of a Stmt without adding any new symbols or re-resolving anything, so this seems to be sufficient.
Added a new helper function to the Symboltable --
defineSingleVarDefStmtwhich -- which defines a single variable
(for instance a compiler-introduced temporary) using a char*
name rather than a Symbol, to remove a few steps in this
common case.  The new variable is returned through a final
parameter.

Turned on the processParameters pass by default.

Fixed the output for Steve's test_record_method2 test.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@1293 3a8e244f-b0f2-452b-bcba-4c88e055c3ca