return_type_info() now correctly return dtVoid for those functions which do not return a value.
In making this change, 2 regressions have occured, both
of which I have make .future and assigned to Brad.

   test/classes/deitz/test_class_with2.future
   test/modules/bradc/manymodules/flatmodules.future

The issue here is that we are trying to generate code for functions
which are never used.  Analysis will tell you the types
for functions which are called, but dtUnknown for things in functions which
are never called.  I can tell you if a function is never called
so that you can avoid generating code.

I have added a new function to analysis.h:

int function_is_used(FnSymbol *fn);

for this purpose.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@1335 3a8e244f-b0f2-452b-bcba-4c88e055c3ca