Removed future tags for these tests, now fixed.
Note that I'd prefer to restrict the .future tag
for use on new files that have never worked, not
as a way of making new regressions seem less bad.
If we need a way of assigning regressions to
specific people, I can implement one.  I didn't
add it at the same time as futures out of laziness,
and because to this point we've had few enough
regressions and developers that it hasn't been
difficult to keep track of which are in whose
court.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@1341 3a8e244f-b0f2-452b-bcba-4c88e055c3ca