Added support for simple type_variables in a form sufficien to handle test_typevar_function{1,2}.chpl to analysis.
I have put a simple implementation of Type::instantiate_generic
in place which is sufficient for this test case.
I have also made a stub for FnSymbol::instantiate_generic
which currently returns NULL.  This causes these
two test cases to SEGV.  The implementation (which
I am leaving to Steve) could be a modification
of FnSymbol::clone() (or perhaps a call to a shared
produced).


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@1382 3a8e244f-b0f2-452b-bcba-4c88e055c3ca