Removed commented out versions of FnDefStmt::clone,etc.
Removed unused currentFn in Symboltable since this was out-of-date and
not used.

Added commented out version FnSymbol::clone which doesn't work because
analysis stores the type information of a cloned FnSymbol in its
FnDefStmt.  Thus the cloned version must create a FnDefStmt even if it
is ignored.  This has to do with the recent changes.

I now see analysis's representation as being more tightly coupled to
the high-level representation than I thought.  This explains much.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@1384 3a8e244f-b0f2-452b-bcba-4c88e055c3ca