Added the following interface parameter to clone:
  Map<BaseAST*,BaseAST*>* map

This results in the following cloning interfaces:

  TypeDefStmt* clone(CloneCallback* clone_callback, Map<BaseAST*,BaseAST*>* map);
  FnSymbol* clone(CloneCallback* clone_callback, Map<BaseAST*,BaseAST*>* map);

Currently, clone_callback is used by analysis but, with this checkin,
map is ready to be used instead.  It contains a mapping of everything
that was copied in the clone call.

Souped up the copying mechanism to create a mapping between everything
that is duplicated if the map parameter is not NULL.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@1400 3a8e244f-b0f2-452b-bcba-4c88e055c3ca