Added initial support for secondary methods, bound functions that are not embedded in a class.
Beefed up the noanalysis flag so that class member accesses can now be
resolved most of the time without analysis and so that functions that
don't return anything don't require analysis to determine this.

Added several test cases of secondary methods.  These are not
supported in analysis yet.  And one of the tests still has problems
without analysis.  Futures are set.

Fixed collect_asts to call collect_symbols.  John: This is the code I
sent you.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@1418 3a8e244f-b0f2-452b-bcba-4c88e055c3ca