Blowing off steam after an afternoon of TAR work, and continuing in the ongoing struggle to remove my name from the future list... :)
Added a new traversal to check the semantics of
the program after analysis has run and we've
incorporated its results into the AST.  Currently,
the only thing that this pass checks for is
assignment to constant expressions.  Added an
isConst() method to the Expr and Symbol class
hierarchies to help with this effort (and renamed
the old isConst field to be isConstant to avoid
conflicts).

Known problems: blank parameter intents that
should be interpreted as const are not classified
correctly by the current isConst() implementation
(a note to that effect was placed in
ParamSymbol::isConst().


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@1434 3a8e244f-b0f2-452b-bcba-4c88e055c3ca