Changed our write() implementation for floats so that floats with 0 fractional bits (e.g., 3.0) would be printed out as 3.0 rather than 3.  Updated test output to reflect this change.
Some notes: in order to do this, I used a pointer
comparison on the formatting string so that we
could tell whether it was the default format (which
we control so it makes sense to do this), or the user's
format (in which case we should do what they request).
Most of our formatting strings are stored as #defines,
which makes the pointer compare impossible, though we
wouldn't want a strcmp, because we'd want the user's
"%g" to behave as a true C "%g".  Thus, I made the format
for writing floats a char* variable rather than a #define.

I didn't do this for all of the format strings because
in other cases (particularly reading), we rely on the
default format value to be a string literal so that we
can glom two of them together when looking for errors.

In retrospect, the correct way to go might be to pass
in NULL as the format string if the user hasn't specified
one, and store all the format strings as #defines, plugging
the appropriate one in if a NULL is passed in.  This would
make detection of the default trivial, and would allow us
to keep using #defines for all format strings.  I didn't
make this change at this time, though.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@1443 3a8e244f-b0f2-452b-bcba-4c88e055c3ca